+++ zeroconf/Cargo.toml
@@ -38,10 +38,10 @@ clap = { version = "4.4.4", features = ["derive"] }
 avahi-sys = "0.10.0"
 
 [target.'cfg(target_vendor = "apple")'.dependencies]
-bonjour-sys = "0.2.0"
+bonjour-sys = { git = "https://github.com/davidcorrigan714/bonjour-sys.git" } 
 
 [target.'cfg(target_vendor = "pc")'.dependencies]
-bonjour-sys = "0.2.0"
+bonjour-sys = { git = "https://github.com/davidcorrigan714/bonjour-sys.git" } 
 
 [package.metadata.docs.rs]
 default-target = "x86_64-unknown-linux-gnu"
