+++ zeroconf/src/avahi/avahi_util.rs
