+++ zeroconf/src/avahi/entry_group.rs
