+++ zeroconf/src/avahi/resolver.rs
