+++ zeroconf/src/avahi/string_list.rs
