+++ zeroconf/src/avahi/txt_record.rs
