+++ zeroconf/src/lib.rs
@@ -216,10 +216,10 @@ pub mod prelude;
 pub mod service;
 pub mod txt_record;
 
+#[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
+pub mod bonjour;
 #[cfg(target_os = "linux")]
 pub mod linux;
-#[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
-pub mod macos;
 
 pub use browser::{ServiceDiscoveredCallback, ServiceDiscovery};
 pub use interface::*;
@@ -231,21 +231,21 @@ pub use service_type::*;
 pub type MdnsBrowser = linux::browser::AvahiMdnsBrowser;
 /// Type alias for the platform-specific mDNS browser implementation
 #[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
-pub type MdnsBrowser = macos::browser::BonjourMdnsBrowser;
+pub type MdnsBrowser = bonjour::browser::BonjourMdnsBrowser;
 
 /// Type alias for the platform-specific mDNS service implementation
 #[cfg(target_os = "linux")]
 pub type MdnsService = linux::service::AvahiMdnsService;
 /// Type alias for the platform-specific mDNS service implementation
 #[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
-pub type MdnsService = macos::service::BonjourMdnsService;
+pub type MdnsService = bonjour::service::BonjourMdnsService;
 
 /// Type alias for the platform-specific structure responsible for polling the mDNS event loop
 #[cfg(target_os = "linux")]
 pub type EventLoop<'a> = linux::event_loop::AvahiEventLoop<'a>;
 /// Type alias for the platform-specific structure responsible for polling the mDNS event loop
 #[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
-pub type EventLoop<'a> = macos::event_loop::BonjourEventLoop<'a>;
+pub type EventLoop<'a> = bonjour::event_loop::BonjourEventLoop<'a>;
 
 /// Type alias for the platform-specific structure responsible for storing and accessing TXT
 /// record data
@@ -254,7 +254,7 @@ pub type TxtRecord = linux::txt_record::AvahiTxtRecord;
 /// Type alias for the platform-specific structure responsible for storing and accessing TXT
 /// record data
 #[cfg(any(target_vendor = "apple", target_vendor = "pc"))]
-pub type TxtRecord = macos::txt_record::BonjourTxtRecord;
+pub type TxtRecord = bonjour::txt_record::BonjourTxtRecord;
 
 /// Result type for this library
 pub type Result<T> = std::result::Result<T, error::Error>;
